Create backup_script.sh
This file automates the backup of a specified directory to a remote server using rsync. It provides a report on the success or failure of the backup operation.
This script uses rsync to synchronize files from the specified source directory to a remote server. The script creates a report file (backup_report.txt) that contains information about the success or failure of the backup operation.